% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{chrome}
\alias{chrome}
\title{Start chrome driver}
\usage{
chrome(type = c("browser", "selenium"), port = 4567L, path = "wd/hub")
}
\arguments{
\item{type}{Run locally (chromedriver only) or as a remote driver
(selenium)}

\item{port}{Port to run on}

\item{path}{base URL path prefix for commands, e.g. wd/hub}
}
\description{
Start chrome driver
}

