% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gecko.R
\name{gecko}
\alias{gecko}
\title{Start gecko driver}
\usage{
gecko(port = 4567L, version = "latest", log = c("fatal", "error", "warn",
  "info", "config", "debug", "trace"))
}
\arguments{
\item{port}{Port to run on}

\item{version}{what version of geckodriver to run. Default = "latest"
which runs the most recent version. To see other version currently
sourced run binman::list_versions("geckodriver")}

\item{log}{Set Gecko log level [values: fatal, error,
warn, info, config, debug, trace]}

\item{path}{base URL path prefix for commands, e.g. wd/hub}
}
\description{
Start gecko driver
}

