% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predl_chrome_for_testing.R
\name{unziptar_dlfiles}
\alias{unziptar_dlfiles}
\title{Unzip/untar files}
\usage{
unziptar_dlfiles(...)
}
\arguments{
\item{...}{Passed into [binman::unziptar_dlfiles].}
}
\value{
The same as [binman::unziptar_dlfiles].
}
\description{
Unzip or untar downloaded files
}
