% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdman-package.r
\docType{package}
\name{wdman}
\alias{wdman}
\title{wdman}
\description{
Webdriver/Selenium Binary Manager
}
\details{
There are a number of binary files associated with the
Webdriver/Selenium project. This package provides functions to download
these binaries and to manage processes involving them.
}
