% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{chrome}
\alias{chrome}
\title{Start chrome driver}
\usage{
chrome(port = 4567L, version = "latest", path = "wd/hub",
  subprocess = TRUE)
}
\arguments{
\item{port}{Port to run on}

\item{version}{what version of chromedriver to run. Default = "latest"
which runs the most recent version. To see other version currently
sourced run binman::list_versions("chromedriver")}

\item{path}{base URL path prefix for commands, e.g. wd/hub}
}
\description{
Start chrome driver
}

