% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gecko.R
\name{gecko}
\alias{gecko}
\title{Start gecko driver}
\usage{
gecko(port = 4567L, version = "latest", log = c("fatal", "error", "warn",
  "info", "config", "debug", "trace"), verbose = TRUE)
}
\arguments{
\item{port}{Port to run on}

\item{version}{what version of geckodriver to run. Default = "latest"
which runs the most recent version. To see other version currently
sourced run binman::list_versions("geckodriver")}

\item{log}{Set Gecko log level [values: fatal, error,
warn, info, config, debug, trace]}

\item{verbose}{If TRUE, include status messages (if any)}
}
\value{
Returns a list with named elements process, output, error and
    stop. process is the output from calling \code{\link{spawn_process}}
    output, error and stop are functions calling
    \code{\link{process_read}}, \code{\link{process_read}} with "stderr"
    pipe and \code{\link{process_kill}}  respectively  on process.
}
\description{
Start gecko driver
}
\examples{
\dontrun{
gDrv <- gecko()
gDrv$output()
gDrv$stop()
}
}

