% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predl_chrome_for_testing.R
\name{unziptar_chromedriver}
\alias{unziptar_chromedriver}
\title{Unzip/untar the chromedriver file}
\usage{
unziptar_chromedriver(...)
}
\arguments{
\item{...}{Passed into [binman::unziptar_dlfiles].}
}
\value{
The same as [binman::unziptar_dlfiles].
}
\description{
Unzip or untar a downloaded chromedriver file, then extract it from
its folder.
}
