% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predl_chrome_for_testing.R
\name{unziptar_dlfiles}
\alias{unziptar_dlfiles}
\title{Unzip/untar downloaded files}
\usage{
unziptar_dlfiles(...)
}
\arguments{
\item{...}{Passed into \link[binman:unziptar_dlfiles]{binman::unziptar_dlfiles}.}
}
\value{
The same as \link[binman:unziptar_dlfiles]{binman::unziptar_dlfiles}: a list of character
vectors indicating files processed.
}
\description{
Unzip or untar downloaded files.  Copies files into the
expected position.
}
